Matched func_80B92660 in ovl_Bg_Dblue_Elevator. Stack was hard to fix.

+++ src/overlays/actors/ovl_Bg_Dblue_Elevator/z_bg_dblue_elevator.h
@@ -26,12 +26,12 @@ typedef struct BgDBlueElevatorStruct1
 {
   s32 unk0;
   BgDBlueElevatorStruct1Func unk4;
-  s32 unk8;
+  f32 unk8;
   s8 unkC;
   s8 unkD;
-  s32 unk10;
-  s32 unk14;
-  s32 unk18;
+  f32 unk10;
+  f32 unk14;
+  f32 unk18;
 } BgDBlueElevatorStruct1;
 
 #endif // Z_BG_DBLUE_ELEVATOR_H
